build(webpack): add package.json to declare ESM module type
